+++ psiturk/experiment.py
@@ -74,12 +74,16 @@ app.secret_key = CONFIG.get('Server Parameters', 'secret_key')
 def check_templates_exist():
     try:
         try_these = ['thanks-mturksubmit.html', 'closepopup.html']
-        [app.jinja_env.get_template(try_this) for try_this in try_these]
+        [ app.jinja_env.get_template(try_this) for try_this in try_these ]
     except TemplateNotFound as e:
-        raise RuntimeError('Missing one of the following templates: {}. Copy these over from a freshly-created psiturk example experiment. {}: {}'.format(', '.join(try_these), type(e).__name__, str(e)))
 
-if not CONFIG['Shell Parameters'].getboolean('use_psiturk_ad_server'):
-    check_templates_exist()
+        raise RuntimeError((
+            f"Missing one of the following templates: {', '.join(try_these)}."
+            f"Copy these over from a freshly-created psiturk example experiment."
+            f"{type(e).__name__, str(e)}"
+        ))
+
+check_templates_exist()
 
 
 # Serving warm, fresh, & sweet custom, user-provided routes
@@ -121,14 +125,14 @@ scheduler = Scheduler(jobstores=jobstores, timezone=utc)
 app.apscheduler = scheduler
 scheduler.app = app
 
-if (os.getenv('PSITURK_DO_SCHEDULER', False)):
+if CONFIG.get('Server Parameters', 'do_scheduler'):
     scheduler.start()
 
 
 #
 # Dashboard
 #
-if (os.getenv('PSITURK_ENABLE_DASHBOARD', False)):
+if CONFIG.get('Server Parameters', 'enable_dashboard'):
     from .dashboard import dashboard, init_app as dashboard_init_app # management dashboard
     app.register_blueprint(dashboard)
     dashboard_init_app(app)
@@ -158,7 +162,7 @@ def handle_exp_error(exception):
     """Handle errors by sending an error page."""
     app.logger.error(
         "%s (%s) %s", exception.value, exception.errornum, str(dict(request.args)))
-    return exception.error_page(request, CONFIG.get('HIT Configuration',
+    return exception.error_page(request, CONFIG.get('Task Parameters',
                                                     'contact_email_on_error'))
 
 # for use with API errors
@@ -191,7 +195,7 @@ def get_random_condcount(mode):
 
     Returns a tuple: (cond, condition)
     """
-    cutofftime = datetime.timedelta(minutes=-CONFIG.getint('Server Parameters',
+    cutofftime = datetime.timedelta(minutes=-CONFIG.getint('Task Parameters',
                                                            'cutoff_time'))
     starttime = datetime.datetime.now(datetime.timezone.utc) + cutofftime
 
@@ -266,7 +270,7 @@ def check_worker_status():
     else:
         worker_id = request.args['workerId']
         assignment_id = request.args['assignmentId']
-        allow_repeats = CONFIG.getboolean('HIT Configuration', 'allow_repeats')
+        allow_repeats = CONFIG.getboolean('Task Parameters', 'allow_repeats')
         if allow_repeats:  # if you allow repeats focus on current worker/assignment combo
             try:
                 part = Participant.query.\
@@ -308,7 +312,7 @@ def advertisement():
     user_agent_string = request.user_agent.string
     user_agent_obj = user_agents.parse(user_agent_string)
     browser_ok = True
-    browser_exclude_rule = CONFIG.get('HIT Configuration', 'browser_exclude_rule')
+    browser_exclude_rule = CONFIG.get('Task Parameters', 'browser_exclude_rule')
     for rule in browser_exclude_rule.split(','):
         myrule = rule.strip()
         if myrule in ["mobile", "tablet", "touchcapable", "pc", "bot"]:
@@ -362,7 +366,7 @@ def advertisement():
     except exc.SQLAlchemyError:
         status = None
 
-    allow_repeats = CONFIG.getboolean('HIT Configuration', 'allow_repeats')
+    allow_repeats = CONFIG.getboolean('Task Parameters', 'allow_repeats')
     if (status == STARTED or status == QUITEARLY) and not debug_mode:
         # Once participants have finished the instructions, we do not allow
         # them to start the task again.
@@ -371,22 +375,17 @@ def advertisement():
         # 'or status == SUBMITTED' because we suspect that sometimes the post
         # to mturk fails after we've set status to SUBMITTED, so really they
         # have not successfully submitted. This gives another chance for the
-        # submit to work when not using the psiturk ad server.
-        use_psiturk_ad_server = CONFIG.getboolean(
-            'Shell Parameters', 'use_psiturk_ad_server')
-        if not use_psiturk_ad_server:
-            # They've finished the experiment but haven't successfully submitted the HIT
-            # yet.
-            return render_template(
-                'thanks-mturksubmit.html',
-                using_sandbox=(mode == "sandbox"),
-                hitid=hit_id,
-                assignmentid=assignment_id,
-                workerid=worker_id
-            )
-        else:
-            # Show them a thanks message and tell them to go away.
-            return render_template('thanks.html')
+        # submit to work.
+
+        # They've finished the experiment but haven't successfully submitted the HIT
+        # yet.
+        return render_template(
+            'thanks-mturksubmit.html',
+            using_sandbox=(mode == "sandbox"),
+            hitid=hit_id,
+            assignmentid=assignment_id,
+            workerid=worker_id
+        )
     elif already_in_db and not (debug_mode or allow_repeats):
         raise ExperimentError('already_did_exp_hit')
     elif status == ALLOCATED or not status or debug_mode:
@@ -429,22 +428,6 @@ def give_consent():
     )
 
 
-def get_ad_via_hitid(hit_id):
-    ''' Get ad via HIT id '''
-    username = CONFIG.get('psiTurk Access', 'psiturk_access_key_id')
-    password = CONFIG.get('psiTurk Access', 'psiturk_secret_access_id')
-    try:
-        req = requests.get('https://api.psiturk.org/api/ad/lookup/' + hit_id,
-                           auth=(username, password))
-    except:
-        raise ExperimentError('api_server_not_reachable')
-    else:
-        if req.status_code == 200:
-            return req.json()['ad_id']
-        else:
-            return "error"
-
-
 @app.route('/exp', methods=['GET'])
 @nocache
 def start_exp():
@@ -468,7 +451,7 @@ def start_exp():
 
     # Check first to see if this hitId or assignmentId exists.  If so, check to
     # see if inExp is set
-    allow_repeats = CONFIG.getboolean('HIT Configuration', 'allow_repeats')
+    allow_repeats = CONFIG.getboolean('Task Parameters', 'allow_repeats')
     if allow_repeats:
         matches = Participant.query.\
             filter(Participant.workerid == worker_id).\
@@ -541,23 +524,7 @@ def start_exp():
             if other_assignment:
                 raise ExperimentError('already_did_exp_hit')
 
-    use_psiturk_ad_server = CONFIG.getboolean(
-        'Shell Parameters', 'use_psiturk_ad_server')
-    if use_psiturk_ad_server and (mode == 'sandbox' or mode == 'live'):
-        # If everything goes ok here relatively safe to assume we can lookup
-        # the ad.
-        ad_id = get_ad_via_hitid(hit_id)
-        if ad_id != "error":
-            if mode == "sandbox":
-                ad_server_location = 'https://sandbox.ad.psiturk.org/complete/'\
-                    + str(ad_id)
-            elif mode == "live":
-                ad_server_location = 'https://ad.psiturk.org/complete/' +\
-                    str(ad_id)
-        else:
-            raise ExperimentError('hit_not_registered_with_ad_server')
-    else:
-        ad_server_location = '/complete'
+    ad_server_location = '/complete'
 
     return render_template(
         'exp.html', uniqueId=part.uniqueid,
@@ -566,7 +533,7 @@ def start_exp():
         adServerLoc=ad_server_location,
         mode=mode,
         contact_address=CONFIG.get(
-            'HIT Configuration', 'contact_email_on_error'),
+            'Task Parameters', 'contact_email_on_error'),
         codeversion=CONFIG.get(
             'Task Parameters', 'experiment_code_version')
     )
@@ -717,7 +684,7 @@ def debug_complete():
             if (mode == 'sandbox' or mode == 'live'):
                 return render_template('closepopup.html')
             else:
-                allow_repeats = CONFIG.getboolean('HIT Configuration', 'allow_repeats')
+                allow_repeats = CONFIG.getboolean('Task Parameters', 'allow_repeats')
                 return render_template('complete.html',
                     allow_repeats=allow_repeats, worker_id=user.workerid)
 
@@ -807,7 +774,7 @@ def run_webserver():
     ''' Run web server '''
     host = CONFIG.get('Server Parameters', 'host')
     port = CONFIG.getint('Server Parameters', 'port')
-    print("Serving on http://{}:{}".format(host, port))
+    print(f"Serving on http://{host}:{port}")
     app.config['TEMPLATES_AUTO_RELOAD'] = True
     app.jinja_env.auto_reload = True
     app.run(debug=True, host=host, port=port)