+++ psiturk/setup_example.py
@@ -20,31 +20,23 @@ else:
 
 GLOBAL_CONFIG_FILE = os.path.expanduser(GLOBAL_CONFIG_PATH)
 EXAMPLE_TARGET = os.path.join(os.curdir, "psiturk-example")
-CONFIG_TARGET = os.path.join(EXAMPLE_TARGET, "config.txt")
 
 
 def setup_example():
     ''' Setup example '''
     if os.path.exists(EXAMPLE_TARGET):
-        print("Error, `psiturk-example` directory already exists.  Please \
+        print("Error, `psiturk-example` directory already exists. Please \
             remove it then re-run the command.")
     else:
         print("Creating new folder `psiturk-example` in the current working \
             directory")
         os.mkdir(EXAMPLE_TARGET)
+
         print("Copying", EXAMPLE_DIR, "to", EXAMPLE_TARGET)
         dir_util.copy_tree(EXAMPLE_DIR, EXAMPLE_TARGET)
-        # change to target directory
-        print("Creating default configuration file (config.txt)")
-        file_util.copy_file(DEFAULT_CONFIG_FILE, CONFIG_TARGET)
-        os.chdir(EXAMPLE_TARGET)
-        os.rename('custom.py.txt', 'custom.py')
 
-        if not os.path.exists(GLOBAL_CONFIG_FILE):
-            print("The following config file does not exist:\n{}\
-                \nCreating default config file at that \
-                location.".format(GLOBAL_CONFIG_FILE))
-            file_util.copy_file(DEFAULT_GLOBAL_CONFIG_FILE, GLOBAL_CONFIG_FILE)
+        # change to target directory
+        os.chdir(EXAMPLE_TARGET)
 
 
 if __name__ == "__main__":